refactor(welcome): rename view to index.vue, prefix section components
welcome-view.vue -> index.vue (thin orchestrator); *-section.vue ->
section-*.vue. Update router import and the e2e login-trigger testid.
