+++ src/router/index.ts
@@ -3,7 +3,7 @@ import { useSessionStore } from '@/stores/session'
 import { useMemberStore } from '@/stores/member'
 import { prefetchMemberDecks } from '@/api/decks'
 import { prefetchMemberById } from '@/api/members'
-import WelcomeView from '@/views/welcome/welcome-view.vue'
+import WelcomeView from '@/views/welcome/index.vue'
 import AuthenticatedView from '@/views/authenticated.vue'
 import PrivacyPolicyView from '@/views/privacy-policy.vue'
 import TermsOfServiceView from '@/views/terms-of-service.vue'
