+++ src/views/welcome/section-pricing.vue
